GdxPayException moved to core project, introduced ItemAlreadyOwnedException
