+++ gdx-pay-android-googleplay/src/main/java/com/badlogic/gdx/pay/android/googleplay/AndroidGooglePlayPurchaseManager.java
@@ -23,6 +23,7 @@ import android.util.Log;
 
 import com.badlogic.gdx.backends.android.AndroidApplication;
 import com.badlogic.gdx.backends.android.AndroidFragmentApplication;
+import com.badlogic.gdx.pay.GdxPayException;
 import com.badlogic.gdx.pay.Information;
 import com.badlogic.gdx.pay.Offer;
 import com.badlogic.gdx.pay.OfferType;
