+++ gdx-pay-android-googleplay/src/main/java/com/badlogic/gdx/pay/android/googleplay/billing/V3GoogleInAppBillingService.java
@@ -20,7 +20,7 @@ import com.badlogic.gdx.pay.Information;
 import com.badlogic.gdx.pay.PurchaseObserver;
 import com.badlogic.gdx.pay.Transaction;
 import com.badlogic.gdx.pay.android.googleplay.ConsumeException;
-import com.badlogic.gdx.pay.android.googleplay.GdxPayException;
+import com.badlogic.gdx.pay.GdxPayException;
 import com.badlogic.gdx.pay.android.googleplay.ResponseCode;
 import com.badlogic.gdx.pay.android.googleplay.billing.converter.PurchaseResponseActivityResultConverter;
