+++ gdx-pay-android-googleplay/src/test/java/com/badlogic/gdx/pay/android/googleplay/AndroidGooglePlayPurchaseManagerTest.java
@@ -6,6 +6,7 @@ import android.content.pm.PackageInfo;
 import android.content.pm.PackageManager;
 
 import com.badlogic.gdx.backends.android.AndroidApplication;
+import com.badlogic.gdx.pay.GdxPayException;
 import com.badlogic.gdx.pay.Information;
 import com.badlogic.gdx.pay.Offer;
 import com.badlogic.gdx.pay.PurchaseObserver;
@@ -39,7 +40,6 @@ import static com.badlogic.gdx.pay.android.googleplay.testdata.TestConstants.PAC
 import static com.badlogic.gdx.pay.android.googleplay.testdata.TransactionObjectMother.transactionFullEditionEuroGooglePlay;
 import static java.util.Collections.singletonList;
 import static java.util.Collections.singletonMap;
-import static org.hamcrest.CoreMatchers.containsString;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertSame;
